+++ cli/commands/watch/tester.ts
@@ -3,6 +3,7 @@ import { readConfig } from "../../mops.js";
 import { ErrorChecker } from "./error-checker.js";
 import { testWithReporter } from "../test/test.js";
 import { SilentReporter } from "../test/reporters/silent-reporter.js";
+import { type ReplicaName } from "../../helpers/deprecate-dfx-replica.js";
 
 export class Tester {
   verbose = false;
@@ -53,12 +54,15 @@ export class Tester {
     this.controller = new AbortController();
 
     let config = readConfig();
+    let replicaType: ReplicaName = config.toolchain?.["pocket-ic"]
+      ? "pocket-ic"
+      : "dfx";
 
     this.currentRun = testWithReporter(
       this.reporter,
       "",
       "interpreter",
-      config.toolchain?.["pocket-ic"] ? "pocket-ic" : "dfx",
+      replicaType,
       true,
       this.controller.signal,
     );